% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize.R
\name{split_dataframe_by_time}
\alias{split_dataframe_by_time}
\title{Split dataframe by time}
\usage{
split_dataframe_by_time(
  dataframe,
  time_column,
  min_len = 0,
  margin = 5,
  rate_rate_change_margin = 0.5,
  ...
)
}
\arguments{
\item{dataframe}{Dataframe that contains time series information in a number of continuous chunks}

\item{time_column}{A column containing time information}

\item{min_len}{A minimum length of a chunk to be outputted}

\item{margin}{Minimum number of local periods between 2 continuous chunks needed to treat them as separate chunks}

\item{rate_rate_change_margin}{Minimum relative change of sample rate between 2 chunks to treat them as separate chunks}

\item{...}{}
}
\value{
A vector of dataframes continuous in time and with constant rate
}
\description{
Divides a dataset into a vector of datasets with continuous time series'.
Devision occurs if the sampling rate changes or suddenly changes values.
}
\examples{
df = data.frame(c(1:10, seq(11,20,2), 1:10),c(1:25))
names(df) = c("time", "data")
split_dataframe_by_time(df, df$time)

[[1]]
time data
1     1    1
2     2    2
3     3    3
4     4    4
5     5    5
6     6    6
7     7    7
8     8    8
9     9    9
10   10   10
[[2]]
time data
11   11   11
12   13   12
13   15   13
14   17   14
15   19   15
[[3]]
time data
16    1   16
17    2   17
18    3   18
19    4   19
20    5   20
21    6   21
22    7   22
23    8   23
24    9   24
25   10   25
}
