% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize.R
\name{vectorize}
\alias{vectorize}
\title{Aggregate windows in a dataframe}
\usage{
vectorize(dataframe, window_size, slide, aggregations = list(), all = c(), ...)
}
\arguments{
\item{dataframe}{Source dataframe with time series data in columns}

\item{window_size}{Size of the window for aggregations}

\item{slide}{Number of rows by which the window will slide. Use the same value as window_size for no intersections and no skipped values}

\item{aggregations}{list with named values. Each element of the list is a vector of strings with function names that will be applied to the column with the same name as the name of the element}

\item{all}{If specified, theese aggregation functions will be applied to all columns}

\item{...}{}
}
\value{
Vectorized dataframe. Name of columns are extended by the aggregation function name. If aggregation function returns multiple values, columns are additionally numbered
}
\description{
Calculates values of a set of aggregating functions for dataframe column values contained in a sliding window.
}
\examples{

df = data.frame(1:12)
names(df) = c("column1")
aggregations <- list(column1=c("mean", "min", "max"))
result = vectorize(df, 3, 3, aggregations)

  column1 mean column1 min column1 max
1            2           1           3
2            5           4           6
3            8           7           9
4           11          10          12
}
